Added the compatibility script
The compatibility script will check if the data that is loaded is compatible with the plugin version, if it detect that it is not, it will automatically try to convert it.

There is a lot of check, and a pre-convertion backup so it should be safe to release in the V2.3.0.

If the compatibility script can't convert it, an error will be written, and the user can just continue with the data like before.